Removed/Renamed HprFactory.
